修正了PMLoader::SMM_NODE_SIZE设置,TreeNode-->FullNode 增加了保护模式代码扇区数 110-->400, 这将避免代码空间不足带来的非常难以调试的bug 所有的在多处引用的常数都被归到一处 修改了tools/bochs run/下的几个程序 修改了Process.h文件的引入方式：现在，它只是Kernel.h的一部分

+++ include/macros/Kernel_macros.h
@@ -22,7 +22,7 @@ KernelSmmWrapper<T>::~KernelSmmWrapper()
 template <class T>
 T* KernelSmmWrapper<T>::getNew()
 {
-	return (T*)Kernel::getTheKernel()->mnewKernel(sizeof(T));
+	return (T*)Kernel::getTheKernel()->mnewKernel((size_t)sizeof(T));
 }
 template <class T>
 void KernelSmmWrapper<T>::withdraw(T *t)
